+++ Changelog.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2024-07-23
+
+### Summary
+
+- Work with framework OriCommJS 1.1.3 and above.
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+[1.1.2]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.1.2
+
 ## [1.1.1] - 2024-07-02
 
 ### Summary
@@ -26,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-[1.1.0]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.1.1
+[1.1.1]: https://github.com/wkloh76/oricommjs_components/releases/tag/1.1.1
 
 ## [1.1.0] - 2024-06-10